third_party/WebKit/LayoutTests/editing/pasteboard/line-feed-between-br-and-b-should-not-reorder-pasted-content.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_not_equals(window.internals, undefined,
        'This test requires window.internals to access clipboard');

    assert_selection(
        [
            '<div contenteditable>',
                '<div>^<br><b>bold<br></b>plain|</div>',
            '</div>',
        ].join(''),
        'cut',
        [
            '<div contenteditable>',
                '|<br>',
            '</div>',
        ].join(''),
        'cut');

    assert_selection(
        [
            '<div contenteditable>',
                '<div>^<br><b>bold<br></b>plain|</div>',
            '</div>',
        ].join(''),
        selection => {
            selection.document.execCommand('cut');
            selection.document.execCommand('paste');
        },
        [
            '<div contenteditable>',
                '<br><b>bold<br></b>plain|',
            '</div>',
        ].join(''),
        'cut and paste');
});
</script>
